text-processor/requirements.txt
torch
transformers
ollama
grpcio
grpcio-tools